fix(web): refine homepage model logos
