+++ apps/web/src/components/landingPage/Home/LatestModelsCarousel.tsx
@@ -54,6 +54,7 @@ export function LatestModelsCarousel({ cards }: LatestModelsCarouselProps) {
 								hideFooterLink
 								metaPlacement="header"
 								providerDateInline
+								titleLines={1}
 								className="h-full rounded-[20px] border-zinc-200 bg-white/90 py-0 shadow-sm ring-1 ring-inset ring-zinc-200/60 hover:translate-y-0 hover:shadow-sm dark:border-zinc-800 dark:bg-zinc-950/80 dark:ring-zinc-800/60"
 							/>
 						</MarqueeItem>
