+++ apps/web/src/components/landingPage/PartnerLogos/providerLogos.ts
@@ -1,0 +1,72 @@
+import { listKnownLogos, resolveLogo } from "@/lib/logos";
+
+const blockedProviderIds = new Set([
+	"openrouter",
+	"huggingface",
+	"hugging-face",
+	"phaseo",
+]);
+
+export const excludedProviderLogos: string[] = [
+	"scira",
+	"t3",
+	"/logos/arxiv.svg",
+	"discord",
+	"x",
+	"github",
+	"instagram",
+	"reddit",
+	"grok",
+	"openrouter",
+	"phaseo",
+	"/social/hugging_face.svg",
+];
+
+function normaliseDescriptor(value: string): string {
+	if (value.startsWith("/")) return value;
+	return value.toLowerCase();
+}
+
+export function getProviderLogos(providerIds: string[]): string[] {
+	const exclusionSet = new Set(
+		excludedProviderLogos.map(normaliseDescriptor)
+	);
+	const knownProviderIds = new Set(
+		providerIds.map((id) => id.trim().toLowerCase()).filter(Boolean)
+	);
+
+	const allLogos = listKnownLogos()
+		.filter(({ id, assets }) => {
+			const normalizedId = String(id).toLowerCase();
+			if (!knownProviderIds.has(normalizedId)) return false;
+			if (blockedProviderIds.has(normalizedId)) return false;
+
+			const hasLanguageAsset = Object.values(assets).some((asset) =>
+				asset?.startsWith("/languages/")
+			);
+			if (hasLanguageAsset) return false;
+			const hasObservabilityAsset = Object.values(assets).some((asset) =>
+				asset?.startsWith("/observability/")
+			);
+			if (hasObservabilityAsset) return false;
+			if (id.startsWith("observability-")) return false;
+
+			const normalisedId = normaliseDescriptor(id);
+			if (exclusionSet.has(normalisedId)) return false;
+			const pathCandidate = normaliseDescriptor(`/logos/${id}.svg`);
+			return !exclusionSet.has(pathCandidate);
+		})
+		.map(({ id }) => id);
+
+	const seen = new Set<string>();
+	const deduped: string[] = [];
+	for (const id of allLogos) {
+		const resolved = resolveLogo(id, { fallbackToColor: false });
+		const src = resolved.src;
+		if (!src || seen.has(src)) continue;
+		seen.add(src);
+		deduped.push(id);
+	}
+
+	return deduped;
+}
